Splitting dna into separate files
